requirements.txt
async-timeout==4.0.3
cachetools==5.5.0
casttube==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
Kivy==2.3.0
Kivy-Garden==0.1.5
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyChromecast==13.1.0
Pygments==2.18.0
pyparsing==3.2.0
pyttsx3==2.98
requests==2.32.3
rsa==4.9
SpeechRecognition==3.11.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zeroconf==0.135.0
